backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
django-cors-headers==4.3.1
shapely==2.0.2
numpy<2
websockets==15.0.1
